File upload download

+++ controllers/admin.js
@@ -1,6 +1,6 @@
 const Product = require ('../models/product');
 const {validationResult} = require ('express-validator/check');
-const mongoose = require('mongoose');
+const FileHelper = require('../util/file');
 
 exports.getAddProduct = (req, res, next) => {
   if (!req.session.isLoggedIn) {
@@ -22,8 +22,25 @@ exports.getAddProduct = (req, res, next) => {
 };
 
 exports.postAddProduct = (req, res, next) => {
-  const {title, imageUrl, price, description} = req.body;
+  const {title, price, description} = req.body;
+  const image = req.file;
   const errors = validationResult (req);
+
+  if(!image){
+    return res.status(422).render ('admin/edit-product', {
+      pageTitle: 'Add Product',
+      path: '/admin/add-product',
+      editing: false,
+      errorMessage: "Attached file is not an image",
+      product: {
+        title: title, 
+        price: price,
+        description: description
+      },
+      validationErrors: []
+    });
+  }
+
   if(!errors.isEmpty()) {
     return res.status(422).render ('admin/edit-product', {
       pageTitle: 'Add Product',
@@ -31,16 +48,17 @@ exports.postAddProduct = (req, res, next) => {
       editing: false,
       errorMessage: errors.array ()[0].msg,
       product: {
-        title: title, 
-        imageUrl: imageUrl, 
+        title: title,
         price: price,
         description: description
       },
       validationErrors: errors.array()
     });
   }
+
+  const imageUrl = image.path;
+
   const product = new Product ({
-    // _id: new mongoose.Types.ObjectId('5cc330facd40d814281c41fb'), //creating product with duplicate id
     title,
     price,
     description,
@@ -54,28 +72,11 @@ exports.postAddProduct = (req, res, next) => {
       res.redirect ('/');
     })
     .catch (err => {
-      // console.log ('An error occured!');
-
-      // render 500 server error page
-          // res.render ('admin/edit-product', {
-          //   pageTitle: 'Add Product',
-          //   path: '/admin/add-product',
-          //   editing: false,
-          //   errorMessage: 'Database operation failed, please try again.',
-          //   product: {
-          //     title: title, 
-          //     imageUrl: imageUrl, 
-          //     price: price,
-          //     description: description
-          //   },
-          //   validationErrors: []
-          // });
-      // res.redirect('/500');
+      console.log ("-1",err);
       const error = new Error(err);
       error.httpStatusCode = 500;
       return next(error);
     });
-  // products.push ({title: title});
 };
 
 exports.getEditProduct = (req, res, next) => {
@@ -99,7 +100,7 @@ exports.getEditProduct = (req, res, next) => {
       });
     })
     .catch (err => {
-      // console.log (err);
+      console.log ("0",err);
       // res.redirect('/500');
       const error = new Error(err);
       error.httpStatusCode = 500;
@@ -108,8 +109,10 @@ exports.getEditProduct = (req, res, next) => {
 };
 
 exports.postEditProduct = (req, res, next) => {
-  const {productId, title, price, imageUrl, description} = req.body;
+  const {productId, title, price, description} = req.body;
+  const image = req.file;
   const errors = validationResult (req);
+
   if(!errors.isEmpty()) {
     return res.status(422).render ('admin/edit-product', {
       pageTitle: 'Edit Product',
@@ -118,7 +121,6 @@ exports.postEditProduct = (req, res, next) => {
       errorMessage: errors.array ()[0].msg,
       product: {
         title: title, 
-        imageUrl: imageUrl, 
         price: price,
         description: description,
         _id: productId
@@ -134,14 +136,18 @@ exports.postEditProduct = (req, res, next) => {
       }
       product.title = title;
       product.price = price;
-      product.imageUrl = imageUrl;
+      if(image) {
+        FileHelper.deleteFile(product.imageUrl);
+        product.imageUrl = image.path;
+      }
       product.description = description;
+      
       product.save ().then (result => {
         res.redirect ('/admin/admins-products-list');
       });
     })
     .catch (err => {
-      // console.log (err)
+      console.log ("1",err);
       const error = new Error(err);
       error.httpStatusCode = 500;
       return next(error);
@@ -150,15 +156,21 @@ exports.postEditProduct = (req, res, next) => {
 
 exports.postdeleteProduct = (req, res, next) => {
   const prodId = req.body.productId;
-  Product.deleteOne ({_id: prodId, userId: req.user._id})
-    .then (() => {
-      res.redirect ('/admin/admins-products-list');
-    })
-    .catch (err => {
-      const error = new Error(err);
-      error.httpStatusCode = 500;
-      return next(error);
-    });
+  Product.findById(prodId).then(product => {
+    if(!product){
+      return next(new Error('Product not found.'))
+    }
+    FileHelper.deleteFile(product.imageUrl);
+    return Product.deleteOne ({_id: prodId, userId: req.user._id});
+  }).then (() => {
+    res.redirect ('/admin/admins-products-list');
+  })
+  .catch (err => {
+    console.log ("2",err);
+    const error = new Error(err);
+    error.httpStatusCode = 500;
+    return next(error);
+  });
 };
 
 exports.getOwnersProductList = (req, res, next) => {
@@ -174,7 +186,7 @@ exports.getOwnersProductList = (req, res, next) => {
       });
     })
     .catch (err => {
-      // console.log (err);
+      console.log ("3",err);
       const error = new Error(err);
       error.httpStatusCode = 500;
       return next(error);
